Documents/workspace/commons/requirements.txt
atomiclong==0.1.1
apscheduler==3.2.0
blosc==1.4.1
boto==2.39.0
boto3==1.3.1
celery>=3.1,<4.0
enum34==1.1.3
eventlet>=0.17.4,<=0.21.0
falcon==1.1.0
influxdb==4.1.0
msgpack-python==0.4.8
newrelic==2.46.0.37
paramiko==2.2.1
psutil==4.2.0
pycrypto==2.6.1
pygerduty==0.35.1
pykafka==2.5.0
pymongo>=2.8,<=3.2.2
redlock==1.2.0
redis==2.10.3
slacker==0.9.24
statsd==3.2.1
uwsgi>=2.0.9